move runtime enums outside of llvm backend
